+++ ksyun/service_ksyun_vpc.go
@@ -3511,9 +3511,6 @@ func (s *VpcService) ReadAndSetVpnTunnels(d *schema.ResourceData, r *schema.Reso
 					return result
 				},
 			},
-			"CreateTime": {
-				Field: "vpn_tunnel_create_time",
-			},
 		},
 	})
 }